Fix incorrect password loop
If when setting the user or admin password you enter the second
passoword incorrectly and can't remember what it was set to originally
you'd enter a loop until you enter the correct password. Updated to put
the user back to the beginning to allow them to enter a password again.